tilefy/requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
flask==2.1.2
Pillow==9.1.1
PyYAML==6.0
redis==4.3.3
requests==2.28.0
uwsgi==2.0.20
